Re-organise folder structure.
