+++ Android/app/src/main/java/com/vigbag/android/application/App.kt
@@ -1,0 +1,16 @@
+package com.vigbag.android.application
+
+import android.app.Application
+
+class App : Application() {
+    
+    override fun onCreate() {
+        super.onCreate()
+        instance = this
+    }
+    
+    companion object {
+        lateinit var instance: App
+    }
+    
+}
