+++ Android/app/src/main/java/com/vigbag/android/model/ChildItemDataClass.kt
@@ -1,4 +1,4 @@
-package com.vigbag.android
+package com.vigbag.android.model
 
 data class ChildItemDataClass(val itemImage:Int,val itemName:String,
                               val itemSpecs:String,val itemRating:Double,val itemReviews:String)
