+++ Android/app/src/main/java/com/vigbag/android/model/ParentItemDataClass.kt
@@ -1,3 +1,3 @@
-package com.vigbag.android
+package com.vigbag.android.model
 
 data class ParentItemDataClass(var title:String,var ChildModelClass:List<ChildItemDataClass>)
