+++ Android/app/src/main/java/com/vigbag/android/ui/login/SignInFragment.kt
@@ -1,4 +1,4 @@
-package com.vigbag.android
+package com.vigbag.android.ui.login
 
 import android.os.Bundle
 import android.view.View
@@ -6,8 +6,9 @@ import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
 import androidx.preference.PreferenceManager
+import com.vigbag.android.R
 import com.vigbag.android.databinding.FragmentSignInBinding
-import com.vigbag.android.util.Constants.preferences
+import com.vigbag.android.util.Constants.Preferences
 
 class SignInFragment : Fragment(R.layout.fragment_sign_in) {
     
@@ -17,8 +18,8 @@ class SignInFragment : Fragment(R.layout.fragment_sign_in) {
     private fun validateCredentials() {
 
         val sharedPreferences = PreferenceManager.getDefaultSharedPreferences(requireContext())
-        val savedEmail = sharedPreferences.getString(preferences.EMAIL, "")
-        val savedPassword = sharedPreferences.getString(preferences.PASSWORD, "")
+        val savedEmail = sharedPreferences.getString(Preferences.EMAIL, "")
+        val savedPassword = sharedPreferences.getString(Preferences.PASSWORD, "")
 
         binding.etxtEmail.setText(savedEmail)
         binding.etxtPassword.setText(savedPassword)
