+++ Android/app/src/main/java/com/vigbag/android/ui/login/SignUpFragment.kt
@@ -1,4 +1,4 @@
-package com.vigbag.android
+package com.vigbag.android.ui.login
 
 import android.content.Context
 import android.os.Bundle
@@ -6,6 +6,7 @@ import android.view.View
 import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
+import com.vigbag.android.R
 import com.vigbag.android.databinding.FragmentSignUpBinding
 
 class SignUpFragment : Fragment(R.layout.fragment_sign_up) {
