+++ Android/app/src/main/java/com/vigbag/android/util/Constants.kt
@@ -2,7 +2,7 @@ package com.vigbag.android.util
 
 object Constants {
     
-    object preferences {
+    object Preferences {
         const val EMAIL = "EMAIL_PREFERENCE"
         const val PASSWORD = "PASSWORD_PREFERENCE"
     }
