+++ Android/app/src/main/res/layout/fragment_home.xml
@@ -8,7 +8,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/ColorPrimarySecondaryVariant"
-    tools:context=".HomeFragment">
+    tools:context=".ui.home.HomeFragment">
 
    <com.google.android.material.card.MaterialCardView
        android:id="@+id/cardView"
