+++ Android/app/src/main/res/layout/fragment_sign_in.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/ColorPrimary"
-    tools:context=".SignInFragment">
+    tools:context=".ui.login.SignInFragment">
 
     <TextView
         android:id="@+id/txt_welcome"
