+++ Android/app/src/main/res/layout/fragment_sign_up.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/ColorPrimary"
-    tools:context=".SignUpFragment">
+    tools:context=".ui.login.SignUpFragment">
 
     <TextView
         android:id="@+id/txt_create_account"
